--- a/acceptance/config_test.go
+++ b/acceptance/config_test.go
@@ -95,7 +95,7 @@ func LoadConfig(t *testing.T, dir string) (TestConfig, string) {
 	configs := FindConfigs(t, dir)
 
 	if len(configs) == 0 {
-		t.Fatalf("No configs found for %s", dir)
+		return TestConfig{}, "(no config)"
 	}
 
 	result := DoLoadConfig(t, configs[0])